Conflicts:
	sources/Hub/Drivers/HubQualityDriverArgs.cs
	sources/Services.Hub/Quality/HubQualityService.cs

+++ sources/Services.Hub/Quality/IHubQualityDriverArgs.cs
@@ -3,6 +3,6 @@
     public interface IHubQualityDriverArgs
     {
         byte DeviceId { get; set; }
-        int Rating { get; set; }
+        byte Rating { get; set; }
     }
 }
